requirements.txt
# Core Dependencies
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.2.0

# Web Interface
streamlit>=1.20.0

# Configuration
pyyaml>=6.0

# Optional: Progress bars
tqdm>=4.65.0
